Move galaxy.tools.parameters.outputs into galaxy.tools.paraser.output_actions.
This module has not external dependencies so it is prefect for the parser module. There still remains work to be done to separate it from XML.
